server/requirements.txt
grpcio
grpcio-tools
pymupdf
pillow
pytesseract
